GenERP/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
billiard==3.6.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.1
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
dj-database-url==0.5.0
Django==4.2
django-allauth==0.54.0
django-autoslug==1.9.9
django-ckeditor==6.5.1
django-countries==7.2.1
django-crispy-forms==1.14.0
django-defender==0.9.7
django-filter==21.1
django-import-export==3.2.0
django-js-asset==2.0.0
django-otp==1.1.6
django-phonenumber-field==7.0.2
django-user-agents==0.4.0
django-zxcvbn-password==2.1.1
docopt==0.6.2
et-xmlfile==1.1.0
ez-setup==0.9
html5lib==1.1
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
isort==5.9.3
lxml==4.9.2
MarkupPy==1.14
msgpack==1.0.5
num2words==0.5.12
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
phonenumbers==8.12.57
Pillow==9.5.0
psutil==5.9.5
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pyotp==2.6.0
pyparsing==2.4.7
pypdf==3.8.1
PyPDF2==1.27.12
pypng==0.20220715.0
pyshorteners==1.0.1
python-bidi==0.4.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
redis==3.5.3
regex==2023.5.5
reportlab==4.0.0
requests==2.28.2
requests-oauthlib==1.3.1
roundrobin==0.0.4
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
svglib==1.5.1
tablib==3.4.0
tinycss2==1.2.1
txaio==21.2.1
typing_extensions==4.5.0
tzdata==2021.5
tzlocal==4.3
ua-parser==0.10.0
uritools==4.0.1
urllib3==1.26.15
urllib3-secure-extra==0.1.0
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.1.2
whitenoise==5.3.0
xhtml2pdf==0.2.10
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
zxcvbn==4.4.28
